backend/requirements.txt
fastapi[standard]
psycopg2
openai
python-dotenv